requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
motor 
pydantic
bson